+++ CHANGES.md
@@ -9,6 +9,24 @@ Note to self: Breaking changes must increment either
 
 -->
 
+## 0.32.0 (2024-07-10)
+
+_**Breaking**_
+
+> No breaking changes were introduced in this version.
+
+_**Features**_
+
+- feat: add validator for `sha384` hash by @msamsami in [#387](https://github.com/python-validators/validators/pull/387)
+
+_**Maintenance**_
+
+- maint: bump version by @msamsami in [#387](https://github.com/python-validators/validators/pull/387)
+
+**Full Changelog**: [`0.31.0...0.32.0`](https://github.com/python-validators/validators/compare/0.31.0...0.32.0)
+
+---
+
 ## 0.31.0 (2024-07-08)
 
 _**Breaking**_